feat(pyq): one reader, no login, no duplicated questions (G006)
G002 left two copies of every question on the paper page: an interactive
one-at-a-time player on top, and a read-only full paper below it. Answering
happened in the player; the questions you could actually read were inert.

Now there is one list. Every question is on the page in order and every
question is answerable — pick an option, press Check Answer, get correct or
incorrect and the explanation — with no account, no modal and no gate
anywhere on the page. The answer key is already in the page, so checking no
longer waits on a network round trip.

Sections render sequentially rather than behind tabs: tabbing would have
hidden two thirds of the paper behind display:none, which is the content
this work exists to expose. The question palette goes with the player —
scrolling replaces it, which also retires the fixed mobile bottom bar and
the clearance bug it caused.

The site header is back on this page. It has never been rendered here,
which barely mattered while the page was walled off but leaves a visitor
arriving from Google with no logo, no navigation and no way to sign in.

Removed as dead: PyqPaperPlayer, PyqPaperSolutions, the per-question
solution route and fetchPyqQuestionSolution, and the PyqSolution type.

Measured: typical paper 124 -> 95 KB gzipped, largest 196 -> 144 KB.
Verified on the build: 68 questions, 68 Check Answer buttons, 68
explanations, all three sections present, nothing display:none, one h1.

Spec: planning/G006-pyq-single-reader.md
